move tests to service package as no integration testing will be done
